# BACK/swagger.yaml
openapi: 3.0.0
info:
  title: API Doceria Gourmet IANES
  description: >
    API RESTful para gerenciamento de estoque da Doceria Gourmet IANES.
    Controla produtos, lotes, movimentações e notificações de validade e estoque baixo.
  version: 1.0.0
  contact:
    name: Sidney Campos Aragão
servers:
  - url: http://localhost:8080/v1/IANES
    description: Ambiente de desenvolvimento
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErroValidacao:
      type: object
      properties:
        status:       { type: string,  example: "Bad Request" }
        status_code:  { type: integer, example: 400 }
        message:      { type: string,  example: "Campos obrigatórios não preenchidos corretamente." }
    ErroNaoEncontrado:
      type: object
      properties:
        status:       { type: string,  example: "Not Found" }
        status_code:  { type: integer, example: 404 }
        message:      { type: string,  example: "Nenhum item encontrado." }
...
paths:
...
  /categoria:
    get:
      tags: [Categoria]
      summary: Listar todas as categorias
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Categorias retornadas com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  categorias:
                    - id_categoria: 1
                      nome_categoria: "Bolo"
                    - id_categoria: 2
                      nome_categoria: "Torta"
                    - id_categoria: 3
                      nome_categoria: "Doce Porção"

    post:
      tags: [Categoria]
      summary: Cadastrar nova categoria
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            example:
              nome_categoria: "Bolo"
      responses:
        '201':
          description: Categoria criada com sucesso
          content:
            application/json:
              example:
                status: "Created"
                status_code: 201
                message: "Item criado com sucesso!"
                response:
                  categoria:
                    id_categoria: 1
                    nome_categoria: "Bolo"
        '400': { description: Nome inválido ou já cadastrado, content: { application/json: { schema: { $ref: '#/components/schemas/ErroValidacao' } } } }
  /categoria/{id}:
    get:
      tags: [Categoria]
      summary: Buscar categoria por ID
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Categoria encontrada
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  categoria:
                    id_categoria: 1
                    nome_categoria: "Bolo"
        '404': { description: Categoria não encontrada, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoEncontrado' } } } }

    put:
      tags: [Categoria]
      summary: Atualizar categoria
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      requestBody:
        required: true
        content:
          application/json:
            example:
              nome_categoria: "Bolo Especial"
      responses:
        '200':
          description: Categoria atualizada com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Item atualizado com sucesso!"
                response:
                  categoria:
                    id_categoria: 1
                    nome_categoria: "Bolo Especial"
        '400': { description: Dados inválidos, content: { application/json: { schema: { $ref: '#/components/schemas/ErroValidacao' } } } }

    delete:
      tags: [Categoria]
      summary: Excluir categoria
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Categoria excluída com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Item deletado com sucesso!"
        '404': { description: Categoria não encontrada, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoEncontrado' } } } }

  # ══════════════════════════════════════════════
  # PRODUTO
  # ══════════════════════════════════════════════
...
  /lote/dashboard:
    get:
      tags: [Lote]
      summary: Dashboard consolidado do estoque em tempo real
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Dashboard retornado com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  dashboard:
                    - id_produto: 1
                      nome_produto: "Bolo de Chocolate"
                      codigo_identificador: 1001
                      nome_categoria: "Bolo"
                      limite_minimo_alerta: 3
                      quantidade_total: 8
                      status_critico: "No prazo"
                      estoque_baixo: false
                      total_lotes: 2
                    - id_produto: 2
                      nome_produto: "Torta de Morango"
                      codigo_identificador: 1002
                      nome_categoria: "Torta"
                      limite_minimo_alerta: 2
                      quantidade_total: 1
                      status_critico: "Alerta"
                      estoque_baixo: true
                      total_lotes: 1
  /lote/alertas:
    get:
      tags: [Lote]
      summary: Produtos com estoque abaixo do mínimo
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Alertas de estoque baixo retornados
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  alertas:
                    - id_produto: 2
                      nome_produto: "Torta de Morango"
                      limite_minimo_alerta: 2
                      quantidade_atual: 1
                  total_alertas: 1
  /lote/status:
    get:
      tags: [Lote]
      summary: Filtrar lotes por status de validade
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: valor
          schema:
            type: string
            enum: [No prazo, Alerta, Vencido]
            example: "Vencido"
      responses:
        '200':
          description: Lotes filtrados retornados
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  lotes:
                    - id_lote: 3
                      data_vencimento: "2026-05-25"
                      status_validade: "Vencido"
                      nome_produto: "Torta de Morango"
                      nome_categoria: "Torta"
  /lote:
    get:
      tags: [Lote]
      summary: Listar todos os lotes
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lotes retornados com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  lotes:
                    - id_lote: 1
                      data_fabricacao: "2026-05-27T08:00:00"
                      data_vencimento: "2026-06-03"
                      tipo_medida: "porcao"
                      quantidade_porcoes: 12
                      quantidade_peso: null
                      status_validade: "No prazo"
                      nome_produto: "Bolo de Chocolate"
                      nome_categoria: "Bolo"

    post:
      tags: [Lote]
      summary: Cadastrar novo lote
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            example:
              id_produto: 1
              data_fabricacao: "2026-05-27T08:00:00"
              data_vencimento: "2026-06-03"
              tipo_medida: "porcao"
              quantidade: 12
      responses:
        '201':
          description: Lote criado com sucesso
          content:
            application/json:
              example:
                status: "Created"
                status_code: 201
                message: "Item criado com sucesso!"
                response:
                  lote:
                    id_lote: 1
                    data_vencimento: "2026-06-03"
                    status_validade: "No prazo"
                    quantidade_porcoes: 12
        '400': { description: Dados inválidos, content: { application/json: { schema: { $ref: '#/components/schemas/ErroValidacao' } } } }
        '404': { description: Produto não encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/ErroNaoEncontrado' } } } }
  /lote/{id}:
    get:
      tags: [Lote]
      summary: Buscar lote por ID
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Lote encontrado
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  lote:
                    id_lote: 1
                    status_validade: "No prazo"
                    nome_produto: "Bolo de Chocolate"

    put:
      tags: [Lote]
      summary: Atualizar lote (recalcula status automaticamente)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      requestBody:
        required: true
        content:
          application/json:
            example:
              data_vencimento: "2026-06-10"
      responses:
        '200':
          description: Lote atualizado com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Item atualizado com sucesso!"
                response:
                  lote:
                    id_lote: 1
                    data_vencimento: "2026-06-10"
                    status_validade: "No prazo"

    delete:
      tags: [Lote]
      summary: Excluir lote (falha se houver movimentações vinculadas)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Lote excluído com sucesso
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Item deletado com sucesso!"
        '500':
          description: Lote possui movimentações vinculadas
          content:
            application/json:
              example:
                status: "Internal Server Error"
                status_code: 500
                message: "Não é possível excluir um lote com movimentações registradas!"

  # ══════════════════════════════════════════════
  # MOVIMENTAÇÃO
  # ══════════════════════════════════════════════
...
  /movimentacao/lote/{id_lote}:
    get:
      tags: [Movimentação]
      summary: Listar movimentações de um lote específico
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id_lote
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Histórico do lote retornado
          content:
            application/json:
              example:
                status: "OK"
                status_code: 200
                message: "Requisição realizada com sucesso!"
                response:
                  movimentacoes:
                    - id_movimentacao: 1
                      tipo_movimentacao: "Entrada"
                      motivo: "Nova fornada"
                      quantidade: 12
                      data_hora: "2026-05-27T10:30:00"
                      nome_usuario: "Sidney Aragão"
                    - id_movimentacao: 2
                      tipo_movimentacao: "Saída"
                      motivo: "Venda balcão"
                      quantidade: 3
                      data_hora: "2026-05-27T14:00:00"
                      nome_usuario: "Sidney Aragão"

  # ══════════════════════════════════════════════
  # NOTIFICAÇÃO
  # ══════════════════════════════════════════════
...